+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 All major and minor version changes will be documented in this file. Details of
 patch-level version changes can be found in [commit messages](../../commits/master).
 
+## 2023.1 - 2023/06/25
+
+- Add `asyncTrace` and `skimageTrace`
+- Raise OSError "svgtrace.trace/ asyncTrace is not supported in Windows Jupyter Notebooks"
+- Documentation: Add Tutorials
+
 ## 2023.0.1 - 2023/03/12
 
 - Raise `FileNotFoundError` when input does not exist